requirements.txt
requests
aiohttp
rich
python-dotenv
